scripts/python/requirements.txt
# - - - - - - - - - - - - - - - - -  General - - - - - - - - - - - - - - - - -
numpy
Pillow
# lama-cleaner>=1.2.2
IOPaint==1.5.3
ultralytics


# - - - - - - - - - - - - - - - - -  waifu2x - - - - - - - - - - - - - - - - -
tqdm
wand >= 0.6.10
fonttools
scipy
waitress
bottle
diskcache
flake8
psutil
pyyaml
onnx
onnxconverter_common
lpips
packaging

# https://github.com/GallagherCommaJack/dctorch
dctorch

# iw3
timm
numba # only for iww3 sbs training
av >= 12.2.0, < 14.0.0
rembg # for --remove-bg
truststore; python_version >= '3.10'

--extra-index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://download.pytorch.org/whl/cu124

# macOS
torch; sys_platform == "darwin"
torchvision; sys_platform == "darwin"

# Linux
torch==2.4.1+cu124; sys_platform == "linux"
torchvision==0.19.1+cu124; sys_platform == "linux"

# Windows
torch==2.4.1+cu124; sys_platform == "win32"
torchvision==0.19.1+cu124; sys_platform == "win32"


# - - - - - - - - - - - - - - - - -  IOPaint - - - - - - - - - - - - - - - - -
opencv-python
diffusers==0.27.2
huggingface_hub==0.25.2
accelerate
peft==0.7.1
transformers>=4.39.1
safetensors
controlnet-aux==0.0.3
uvicorn
# python-multipart
# python-socketio==5.7.2
typer
pydantic>=2.5.2
rich
loguru
yacs
piexif==1.1.3
omegaconf
easydict
gradio==4.21.0
typer-config==1.4.0
